--- src/utilities/data.js
@@ -1,31 +1,0 @@
-const people = [
-  {
-    id: 1,
-    image:
-      'http://res.cloudinary.com/dvrm4erwh/image/upload/v1692631596/Avatars/Image_26_gpq4zy.jpg',
-    name: 'Olena Doe',
-    raiting: '4',
-    quote:
-      'GooseTrack is impressive, the calendar view and filter options make it easy to stay organized and focused. Highly recommended.',
-  },
-  {
-    id: 2,
-    image:
-      'http://res.cloudinary.com/dvrm4erwh/image/upload/v1692631596/Avatars/Image_26_gpq4zy.jpg',
-    name: 'Alexander Hubbard',
-    raiting: '4',
-    quote:
-      'GooseTrack is incredibly helpful, the sidebar with account management, calendar, and filter options make navigation seamless. Great for staying organized.',
-  },
-  {
-    id: 3,
-    image:
-      'http://res.cloudinary.com/dvrm4erwh/image/upload/v1692631596/Avatars/Image_26_gpq4zy.jpg',
-    name: 'Alexander Hubbard',
-    raiting: '4',
-    quote:
-      'GooseTrack is incredibly helpful, the sidebar with account management, calendar, and filter options make navigation seamless. Great for staying organized.',
-  },
-];
-
-export default people;
