+++ src/utilities/index.js
@@ -1,0 +1,1 @@
+export { getReviews } from './getReviews';
